requirements.txt
Flask
python-chess
gunicorn
flake8
black
pytest